Server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pydantic
pandas
openpyxl
python-multipart
requests